🏗️ Speech aws upload working

+++ informant/utils/s3_manager.py
@@ -29,10 +29,21 @@ class _S3Manager:
         """Uploads a file to AWS S3
         :param: file_origin - where the file come from, file path or url
         :param: file_name - where the file will be located and its name. doesn't need './'"""
+        already_exists = self.check_if_file_exists(file_name)
+        if already_exists:
+            print("File already uploaded")
+            return
+        
         result = self.bucket.upload_file(file_origin, file_name)
         self.last_file_path = file_name
         return file_name
-
+    
+    def check_if_file_exists(self, file_to_search: str):
+        for file in self.bucket.objects.all():
+            if file.key == file_to_search:
+                return True
+        return False
+        
     def delete(self,file_path):
         """Delets a file from AWS S3,
         :param: folder and file name where its located. Doesn't need './' """
@@ -42,6 +53,10 @@ class _S3Manager:
     def get_last_uploaded_file_url(self, file_path: str = None):
         if file_path is None:
             file_path = self.last_file_path
+        already_exists = self.check_if_file_exists(file_path)
+        if already_exists:
+            print("File already uploaded")
+            return
 
         return f"https://{self.bucket_name}.s3.{self.region}.amazonaws.com/{self.last_file_path}"
